Update some files

+++ delete.php
@@ -1,10 +1,20 @@
 <?php
 include "config.php";
-$id= $_GET['id'];
+
+$id= isset($_GET['id']) ? $_GET['id'] : null;
 
 if (!isset($id) || !is_numeric($id))
 {
 die("invalid student id");
+}
+
+$stmt = $connection->prepare("SELECT ID FROM students WHERE id = '$id'");
+$stmt->execute();
+
+$student = $stmt->fetch(PDO::FETCH_ASSOC);
+if (empty($student))
+{
+    die("enter Valid student id");
 }
 
 $stmt = $connection->prepare("DELETE FROM students WHERE id = '$id'");
@@ -14,4 +24,3 @@ header("location:index.php");
 exit;
 ?>
 
-.bu
